refactor: TokenType 패키지 이동
