+++ src/main/java/com/example/solidconnection/auth/domain/TokenType.java
@@ -1,13 +1,13 @@
-package com.example.solidconnection.config.token;
+package com.example.solidconnection.auth.domain;
 
 import lombok.Getter;
 
 @Getter
 public enum TokenType {
 
-    ACCESS("", 1000 * 60 * 60), // 1hour
-    REFRESH("refresh:", 1000 * 60 * 60 * 24 * 7), // 7days
-    KAKAO_OAUTH("kakao:", 1000 * 60 * 60); // 1hour
+    ACCESS("ACCESS:", 1000 * 60 * 60), // 1hour
+    REFRESH("REFRESH:", 1000 * 60 * 60 * 24 * 7), // 7days
+    KAKAO_OAUTH("KAKAO:", 1000 * 60 * 60); // 1hour
 
     private final String prefix;
     private final int expireTime;
