+++ src/test/java/com/example/solidconnection/config/token/TokenProviderTest.java
@@ -1,5 +1,8 @@
 package com.example.solidconnection.config.token;
 
+import com.example.solidconnection.auth.domain.TokenType;
+import com.example.solidconnection.auth.service.TokenProvider;
+import com.example.solidconnection.config.security.JwtProperties;
 import com.example.solidconnection.custom.exception.CustomException;
 import com.example.solidconnection.custom.exception.ErrorCode;
 import com.example.solidconnection.support.TestContainerSpringBootTest;
