+++ src/test/java/com/example/solidconnection/e2e/UniversityDetailTest.java
@@ -1,7 +1,7 @@
 package com.example.solidconnection.e2e;
 
 import com.example.solidconnection.config.token.TokenProvider;
-import com.example.solidconnection.config.token.TokenType;
+import com.example.solidconnection.auth.domain.TokenType;
 import com.example.solidconnection.siteuser.domain.SiteUser;
 import com.example.solidconnection.siteuser.repository.SiteUserRepository;
 import com.example.solidconnection.university.dto.LanguageRequirementResponse;
